==> 00_execution_hub.ipynb <==
import nbformat
from nbconvert.preprocessors import ExecutePreprocessor as EP

# import os
# os.listdir()

notebooks = ['01_static_data_preparation.ipynb',
             '02_infection_data_preparation.ipynb',
             '03_mobility_data_preparation.ipynb',
             '04_mobility_feature_creation.ipynb',
             '05_target_and_prediction.ipynb']

ep_ = EP(timeout=200)

for nb_file in notebooks:
    with open(nb_file) as f:
        nb = nbformat.read(f, nbformat.NO_CONVERT)
        try:
            ep_.preprocess(nb)
        except:
            print('Execuction Error in' nb_file)
